Issue #154 refactor: Mappers 리팩토링
