+++ app/src/main/java/com/lateinit/rightweight/data/repository/impl/UserRepositoryImpl.kt
@@ -6,8 +6,8 @@ import com.lateinit.rightweight.data.datasource.remote.UserRemoteDataSource
 import com.lateinit.rightweight.data.model.local.User
 import com.lateinit.rightweight.data.model.remote.WriteModelData
 import com.lateinit.rightweight.data.repository.UserRepository
+import com.lateinit.rightweight.ui.mapper.toHistoryUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryUiModel
-import com.lateinit.rightweight.util.toHistoryUiModel
 import kotlinx.coroutines.flow.Flow
 import java.time.LocalDate
 import java.time.LocalDateTime
