+++ app/src/main/java/com/lateinit/rightweight/ui/MainViewModel.kt
@@ -4,25 +4,25 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.lateinit.rightweight.data.database.intermediate.DayWithExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
+import com.lateinit.rightweight.data.mapper.remote.toDayField
+import com.lateinit.rightweight.data.mapper.remote.toExerciseField
+import com.lateinit.rightweight.data.mapper.remote.toExerciseSetField
+import com.lateinit.rightweight.data.mapper.remote.toHistoryExerciseField
+import com.lateinit.rightweight.data.mapper.remote.toHistoryExerciseSetField
+import com.lateinit.rightweight.data.mapper.remote.toHistoryField
+import com.lateinit.rightweight.data.mapper.remote.toRoutineField
 import com.lateinit.rightweight.data.model.remote.UpdateData
 import com.lateinit.rightweight.data.model.remote.WriteModelData
 import com.lateinit.rightweight.data.repository.LoginRepository
 import com.lateinit.rightweight.data.repository.UserRepository
 import com.lateinit.rightweight.ui.login.NetworkState
+import com.lateinit.rightweight.ui.mapper.toDayUiModel
+import com.lateinit.rightweight.ui.mapper.toRoutineUiModel
 import com.lateinit.rightweight.ui.model.routine.ExerciseSetUiModel
 import com.lateinit.rightweight.ui.model.routine.ExerciseUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryExerciseSetUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryExerciseUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryUiModel
-import com.lateinit.rightweight.util.toDayField
-import com.lateinit.rightweight.util.toDayUiModel
-import com.lateinit.rightweight.util.toExerciseField
-import com.lateinit.rightweight.util.toExerciseSetField
-import com.lateinit.rightweight.util.toHistoryExerciseField
-import com.lateinit.rightweight.util.toHistoryExerciseSetField
-import com.lateinit.rightweight.util.toHistoryField
-import com.lateinit.rightweight.util.toRoutineField
-import com.lateinit.rightweight.util.toRoutineUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.flow.MutableSharedFlow
