+++ app/src/main/java/com/lateinit/rightweight/ui/calendar/CalendarViewModel.kt
@@ -6,13 +6,13 @@ import com.lateinit.rightweight.data.database.intermediate.HistoryWithHistoryExe
 import com.lateinit.rightweight.data.repository.HistoryRepository
 import com.lateinit.rightweight.data.repository.RoutineRepository
 import com.lateinit.rightweight.data.repository.UserRepository
+import com.lateinit.rightweight.ui.mapper.toDayUiModel
+import com.lateinit.rightweight.ui.mapper.toHistoryUiModel
+import com.lateinit.rightweight.ui.mapper.toRoutineUiModel
 import com.lateinit.rightweight.ui.model.routine.DayUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryUiModel
 import com.lateinit.rightweight.ui.model.ParentDayUiModel
 import com.lateinit.rightweight.ui.model.routine.RoutineUiModel
-import com.lateinit.rightweight.util.toDayUiModel
-import com.lateinit.rightweight.util.toHistoryUiModel
-import com.lateinit.rightweight.util.toRoutineUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
