+++ app/src/main/java/com/lateinit/rightweight/ui/exercise/ExerciseViewModel.kt
@@ -3,9 +3,9 @@ package com.lateinit.rightweight.ui.exercise
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.lateinit.rightweight.data.repository.HistoryRepository
+import com.lateinit.rightweight.ui.mapper.toHistoryUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryExerciseSetUiModel
 import com.lateinit.rightweight.ui.model.history.HistoryExerciseUiModel
-import com.lateinit.rightweight.util.toHistoryUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.map
